updated the merge_text node and moved some nodes around
