+++ nodes/griptape_nodes_library/text/__init__.py
@@ -1,0 +1,1 @@
+"""Text nodes package."""
