+++ nodes/griptape_nodes_library/text/load_text.py
